requirements.txt
fastapi
uvicorn[standard]
openai
pydantic
python-dotenv
requests
websockets
python-multipart
pypdf
jinja2
langchain
langgraph
langchain-openai
